git-svn-id: https://rtnet.svn.sourceforge.net/svnroot/rtnet/trunk@668 be94bdbd-1fe7-0310-aeeb-c95d3c307607

+++ rtnet/rtai_rtdm/rtdm_driver.h
@@ -118,7 +118,7 @@ struct rtdm_operations {
 };
 
 struct rtdm_dev_context {
-    int                         context_flags;
+    unsigned long               context_flags;
     int                         fd;
     atomic_t                    close_lock_count;
     struct rtdm_operations      *ops;
